+++ app/src/main/AndroidManifest.xml
@@ -2,8 +2,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
           xmlns:tools="http://schemas.android.com/tools">
 
-    <uses-permission android:name="android.permission.INTERNET" />
-
     <uses-permission android:name="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
                      tools:ignore="ProtectedPermissions" />
 
@@ -52,7 +50,7 @@
         </activity>
 
         <service
-                android:name=".NotificationListener"
+                android:name=".MediaPlayerService"
                 android:label="@string/notification_listener_service_label"
                 android:permission="android.permission.BIND_NOTIFICATION_LISTENER_SERVICE"
                 android:exported="true">
